requirements.txt
aiogram>=3.0
python-dotenv>=1.0.1
jinja2>=3.1.3
apscheduler>=3.10.4
weasyprint==52.5
python-docx>=1.1.0
beautifulsoup4>=4.12.0